# Break TableReader MultiGet into filter and lookup stages