This PR is the first step in enhancing the coroutines MultiGet to be able to lookup a batch in parallel across levels. By having a separate TableReader function for probing the bloom filters, we can quickly figure out which overlapping keys from a batch are definitely not in the file and can move on to the next level.

Test plan:
1. Add a new unit test
2. Run benchmark to check for regression - 
Without this PR - ```multireadrandom :       9.901 micros/op 807943 ops/sec 60.009 seconds 48483936 operations;  419.2 MB/s (6009992 of 6009992 found)```
With this PR - ```multireadrandom :       9.907 micros/op 807466 ops/sec 60.007 seconds 48453936 operations;  418.9 MB/s (5850992 of 5850992 found) ```